refactor: simplify type dependencies
