+++ src/parser/types.ts
@@ -3,13 +3,7 @@ import type { DxfBlock } from './blocks/types';
 import type { CommonDxfEntity } from './entities/shared';
 import type { CommonDXFObject } from './objects/common';
 import type { DxfObject } from './objects/types';
-import type { BlockRecordTableEntry } from './tables/blockRecord';
-import type { DimStylesTableEntry } from './tables/dimStyle';
-import type { LayerTableEntry } from './tables/layer';
-import type { LTypeTableEntry } from './tables/ltype';
-import type { StyleTableEntry } from './tables/style';
-import type { DxfTable } from './tables/types';
-import type { VPortTableEntry } from './tables/vport';
+import type { BlockRecordTableEntry, DimStylesTableEntry, LayerTableEntry, LTypeTableEntry, StyleTableEntry, DxfTable, VPortTableEntry } from './tables';
 
 export interface ParsedDxf {
     header: DxfHeader;
